refactor password change fields into a reusable component
